+++ diesel/src/lib.rs
@@ -354,3 +354,7 @@ pub use query_builder::debug_query;
 pub use query_builder::functions::{delete, insert_into, insert_or_ignore_into, replace_into,
                                    select, sql_query, update};
 pub use result::Error::NotFound;
+
+pub(crate) mod diesel {
+    pub use super::*;
+}