+++ diesel/src/macros/mod.rs
@@ -1042,15 +1042,6 @@ macro_rules! allow_tables_to_appear_in_same_query {
     () => {};
 }
 
-#[macro_export]
-#[doc(hidden)]
-/// Used by `diesel_derives`, which can't access `$crate`
-macro_rules! __diesel_use_everything {
-    () => {
-        pub use $crate::*;
-    };
-}
-
 /// Gets the value out of an option, or returns an error.
 ///
 /// This is used by `FromSql` implementations.