+++ diesel_derives/src/as_changeset.rs
@@ -61,8 +61,8 @@ pub fn derive(item: syn::DeriveInput) -> Result<quote::Tokens, Diagnostic> {
     Ok(wrap_in_dummy_mod(
         model.dummy_mod_name("as_changeset"),
         quote!(
-            use self::diesel::query_builder::AsChangeset;
-            use self::diesel::prelude::*;
+            use diesel::query_builder::AsChangeset;
+            use diesel::prelude::*;
 
             impl #impl_generics AsChangeset for &'update #struct_name #ty_generics
             #where_clause